requirements.txt
fastapi
uvicorn
jinja2
openai
python-multipart
python-dotenv
qdrant-client
langchain
pymupdf
reportlab
